split into separate lambdas for apple and google
